requirements.txt
absl-py==2.2.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
cloudinary==1.41.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-database-url==2.2.0
Django==5.0.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
Jinja2==3.1.6
keras==3.9.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opt_einsum==3.4.0
optree==0.14.1
packaging==24.1
pillow==10.4.0
protobuf==5.29.4
psycopg==3.1.16
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.0.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
triton==3.2.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.1.3
whitenoise==6.7.0
wrapt==1.17.2
zope.interface==7.2
